Make pvsamv1 into a library
